requirements.txt
dj-database-url~=0.5.0
Django~=3.2.20
django-ajax-selects~=2.0.0
django-auth-ldap~=2.2.0
django-countries~=7.5.1
django-dotenv~=1.4.2
django-filter~=23.2
django-getenv~=1.3.2
django-bootstrap4~=2.3.1
djangorestframework~=3.12.1
djangorestframework-csv~=2.1.0
pep8~=1.7.1
requests~=2.24.0
psycopg2~=2.8.6
git+https://github.com/Work4Labs/django-test-pep8
